+++ packages/auth-client/src/actions/app/status.ts
@@ -20,8 +20,21 @@ export interface StatusAPIResponse {
   bio?: string;
   displayName?: string;
   pfpUrl?: string;
-  verifications?: Hex[];
+  verifications?: string[];
   custody?: Hex;
+  signatureParams: {
+    siweUri: string;
+    domain: string;
+    nonce?: string;
+    notBefore?: string;
+    expirationTime?: string;
+    requestId?: string;
+    redirectUrl?: string;
+  };
+  metadata: {
+    ip: string;
+    userAgent: string;
+  };
 }
 
 const path = "channel/status";